requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Azure Services
azure-storage-blob==12.19.0
openai==1.12.0

# Document Generation
python-docx==1.1.0
weasyprint==60.2
Pillow==10.2.0

# Utilities
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2024.1
